+++ CHANGELOG.md
@@ -16,6 +16,7 @@ This is part of the tergite release v2024.03 that updates the full pipeline for 
 - Changed the way discriminators are loaded to load from the database
 - Removed hard-coded discriminators
 - Upgrade to Python 3.9
+- Removed Labber in job processing and calibration
 
 ### Fixed